inference/pyproject.toml
[tool.poetry]
name = "inference"
version = "0.1.0"
description = ""
authors = ["JChien <jeffchien13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.1"
pandas = "^1.4.3"
scikit-learn = "^1.1.1"
onnxruntime = "^1.12.1"
grpcio = "^1.48.0"
protobuf = "3.19.4"

[tool.poetry.dev-dependencies]
watchdog = { extras = ["watchmedo"], version = "^2.1.9" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
